Extract live helper functions into own module
...and add AdpoptopossWeb.LiveComponent module. This allows to have a
better abstracted approach to pull in shared functions into LiveViews
or LiveComponents only, or use them in both.
